+++ generic/tclEnv.c
@@ -441,12 +441,12 @@ TclUnsetEnv(
 
 #if defined(_WIN32)
     string = Tcl_Alloc(length + 2);
-    memcpy(string, name, (size_t) length);
+    memcpy(string, name, length);
     string[length] = '=';
     string[length+1] = '\0';
 #else
     string = Tcl_Alloc(length + 1);
-    memcpy(string, name, (size_t) length);
+    memcpy(string, name, length);
     string[length] = '\0';
 #endif /* _WIN32 */
 
@@ -691,7 +691,7 @@ ReplaceString(
 		(env.cacheSize + growth) * sizeof(char *));
 	env.cache[env.cacheSize] = newStr;
 	(void) memset(env.cache+env.cacheSize+1, 0,
-		(size_t) (growth-1) * sizeof(char *));
+		(growth-1) * sizeof(char *));
 	env.cacheSize += growth;
     }
 }
